Doc: Document how to use static plugins with CMake
Also, update the existing QTPLUGIN example to import a plugin that is
not imported by default.  This is closer what one would do in reality.

Make the CMake examples match the qmake ones.

Move text that describes CMake and qmake to the start of the section,
and add a subsection for CMake and qmake each.

Fixes: QTBUG-91413
Change-Id: Ie67518dc8757c69f382067a7225c72ff0b512ff8
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 456b4078252e615e35ae8c4157f1e2c645aeace0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>