Fix alarms demo to run successfully
Previously it failed with

 QQmlApplicationEngine failed to load component
 qrc:/alarms/main.qml: No such file or directory

main.cpp was adjusted to use alarms/main.cpp but the resource file
locations were not changed.
Change them to match the expected path in main.cpp.

qtquickcontrols2.conf needs to stay in the root of the resource path,
so it's added via a separate resource.

Amends 302121874ea6e664ea2a59a07bff8fa06983c335

Pick-to: 5.15 6.2 6.3
Fixes: QTBUG-100434
Change-Id: I3436e07600ca7dad8bba45274ba49dc6c137a938
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>