Remove "Generic Linux (x86, x64)" as supported desktop platform
It makes little sense to claim support for a 'Generic Linux' in the
same vein as the distributions we actually do have in the CI, RTA etc.

This also removes the mentioning of native Linux x86 builds, which we
currently don't test for at all.

[ChangeLog][Platform-Specific Changes] Linux x86 (Desktop) was removed
from the list of officially supported platforms. We only provide binaries
for Linux x64 since a while, and also do not actively test native builds
on x86 Linux anymore.

Pick-to: 6.3.0 6.3 6.2
Change-Id: I3d35ac4cb9ab8c7e6b985696ab0fd835f3ddd4fb
Reviewed-by: Lars Knoll <lars.knoll@qt.io>