Fix up the introductory section of What's New in 6.4
The "nothing to see here" line is now spurious as there is something
to see there. Reflowed the \li before it and adjusted the preceding
line to accommodate the fact that there are now two of them. Made
clear that QtSpeech's last appearance was in Qt 5.

Change-Id: Ia32972e560b81892a6c6abfaa43e608145df87b4
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit d95dd62248aca643cf12bd48e0c1628341e39df6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>