Plugins: Update information on deploying Qt plugins
Fix description of incompatibilities; Qt will load plugins
built with a newer Qt version, if it is limited to the patch level.

Remove outdated advise that plugins are only built in release mode,
for a debug_and_release build; they are not.

Also update Qt version numbers in the given example setup.

Pick-to: 6.4
Change-Id: I6d58b43f791371d21d2cd9fe3e8c66fb37d892e4
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>