Fix path for YouTube video thumbnail
A relative path was used to refer to a thumbnail image for a video link
on Qt for Android overview page. While QDoc was able to locate and copy
the image to the output directory, the relative path appears as-is in
the generated .qhp file and is no longer correct, causing qhelpgenerator
to ignore it.

Consolidate all 'extra images' into the same source directory and handle
them in one location in the .qdocconf file.

Pick-to: 6.4
Change-Id: I66531da883bac57fcf32e75c11eb087a648d9798
Reviewed-by: Luca Di Sera <luca.disera@qt.io>