Doc: Fix \qt5 macro usage
The macro attempted to pass the result of \c command as an argument.
This is not supported; while it would be possible to pass an
escaped version of the command, remove the \c styling from the links
completely as that makes the link style consistent with the rest of
the documentation.

Pick-to: 6.4
Change-Id: I766530b1cf0642667bb1a24782ce752f5b63f0d9
Reviewed-by: Paul Wicking <paul.wicking@qt.io>