Doc: Fix \qt5 macro usage
The macro attempted to pass the result of \c command as an argument.
This is not supported; while it would be possible to pass an
escaped version of the command, remove the \c styling from the links
completely as that makes the link style consistent with the rest of
the documentation.

Change-Id: I766530b1cf0642667bb1a24782ce752f5b63f0d9
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 694cb2cb98dcb22c8e01ba4092e40c22ca4c3373)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>