qml/i18n: Mention LanguageChange event for C++ types
When QQmlEngine::retranslate refreshed all bindings, that conveniently
worked also for C++ properties. However, that is no longer the case
since Qt 6.2, which only refreshes actual translation bindings.
As bindings cannot know that a C++ setter uses a translation
macro/function, there needs to be a way to inform the engine about them.

Promote event handling as the recommended approach in
C++ types, which aligns with what needs to be done in a Widgets
application.

Task-number: QTBUG-102393
Pick-to: 6.5
Change-Id: I16b91fcea35331d300fd70756de8819740b5dcb4
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>