Docs: Add tier support information for embedded platforms
Took the table in the share point and created qdoc tables from anything
that had a number in the tier column.
Took the descriptions from the MS teams discussion about what the tiers
mean.
Added a reference to the types of support in the T&Cs.
Added mention of Qt for MCU in the RTOS section, its tier information
cannot be listed here as it is not Qt version dependent.
Updated the relevant pages of the QNX and INTEGRITY documentation to
point to this new location.

Fixes: QTBUG-111489
Pick-to: 6.5 6.5.0
Change-Id: I492cbbcd73ed6728386b182fb920e68f2891f6a7
Reviewed-by: Kai Köhne <kai.koehne@qt.io>