Raise the minimum supported Android version from 6.0 to 8.0, API 23 to 26
While changes in CI are not being made at this stage, we need to move
forward with setting user's expectations that they can no longer rely on
Qt remains compatible with Android versions that are no longer
maintained by Google.

Note: Though we are still supporting Android 8 and 9 at this stage,these
are no longer receiving security patches from Google.

Replaced version references in other parts of the documentation with a
link to this section or removed them entirely where no longer needed.

Pick-to: 6.5 6.5.0
Fixes: QTBUG-111919
Change-Id: I350ee5d518d42ed0877bd09a7da4810a5373d058
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>