Docs: Change armv7a to armeabi-v7a
When trying to configure a Qt for Android kit for armv7a ABI, if using
the name as listed on
https://doc.qt.io/qt-6/android-building.html#architectures "armv7a"
it will result in an "unknown Android ABI" error. The actual name you
need to pass to configure for android-abis is "armeabi-v7a".

Fixes: QTBUG-111914
Change-Id: I3f185fa18eaab25a635e2f554a1b8e88c15a8ce4
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 295a249546ead85a856dde9e917df5ceb4620aa2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>