Doc: Use the \examplecategory macro to tag examples
The '\meta category' command was used for tagging examples with a
specific category, used in Qt Creators Welcome mode.

As we want to also generate lists of examples belonging to a category
elsewhere in the documentation, replace the command with a macro that
expands to the original \meta command and also adds the example to
a group using the \ingroup command. This way, the category names can
be used as arguments to the \generatelist or \annotatedlist commands.

Task-number: QTBUG-112731
Change-Id: I8590894bdd31d69c38a60cdd42f7301e2cecd2ed
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit b597ecb9543c1bd8253baa2c5819b943f9b8ed56)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>