Doc: Replace raw command with section1
QDoc provides the "\raw" command to allow user to inject code in the
generated output that sidesteps QDoc generation rules.

The "\raw" command is intended for special cases that QDoc currently
does not support. Furthermore, the "\raw" command allows for
format-specific customization.

While the "\raw" command is essential for those special cases where QDoc
automatic generation does not satisfy certain constraints, it is
intended to be avoided when QDoc provides a native way of achieving the
same effect.

"doc/src/legal/licenses.qdoc" used a "\raw" command to generate a
level-2 heading, which is equivalent to QDoc's "\section1" command.

The "\raw" command is now replaced with the native QDoc equivalent.

Change-Id: Ifeba63ba7aaed383bf093eaec4e3e819ef6b703e
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>