Fix cmake install TARGETS statement in todolist demo
The TARGETS statement was given ${CMAKE_PROJECT_NAME} as an argument.
That works in a stand alone build of the demo. If the demo is compiled
as part of a toplevel or qtdoc build, the project name will be "Qt".
That doesn't exist and breaks the build.

This patch changes the argument to "ToDoListApp".

Fixes: QTBUG-114592
Change-Id: Ic5311718cbb6250175355e60b0b6eebaf1165e8b
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit 5c2d5701b5fc2ee5a6ad931a06a89b97c08f173a)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>