Rename mediaplayer demo/example to MediaPlayerApp
The same example exists in qtmultimedia, and it is using a similar
target name as well, which can breaks the top-level build, so we need a
new name for it. I thought since this is a demo project, and it's an
app, we can name it like the ToDoListApp.

Change-Id: Ia91786f7a32e1263157091b10683cd5294c292a2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit ee0cdb359c4d6c71a07d5a649b9337512286aae1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>