FxMaterialShowroomApp: Fix broken qmlcomponents file
CMAKE_PROJECT_NAME is used in the qmplcomponents file, referenced by
CMakeLists.txt. This is bad practice for Qt examples, which works only
if the example is built standalone. If it is built as a part of a
submodule with examples, it expands into "Qt". This doesn't exist and
breaks the build.

This patch replaces the variable with FxMaterialShowroomApp.

Pick-to: 6.6 6.5.2
Change-Id: I1aad1218d44103ae6227f064121be1ea2a523ebe
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>