Examples: Check for availability of Quick3D at configure time
The examples do not need to link against Quick3D, but they use Quick3D
QML import. Qt Quick3D is an optional component in the online installer
though, so its easy to miss. By adding Quick3D to find_package(),
we exchange a runtime error with a configure error in this case:

  error: Found package configuration file: C:/Qt/6.6.0/msvc2019_64/lib/cmake/Qt6/Qt6Config.cmake but it set Qt6_FOUND to FALSE so package "Qt6" is considered to be NOT FOUND.
  Reason given by package: Failed to find required Qt component "Quick3D". Expected Config file at "C:/Qt/6.6.0/msvc2019_64/lib/cmake/Qt6Quick3D/Qt6Quick3DConfig.cmake" does NOT exist

Pick-to: 6.5 6.6
Task-number: QTBUG-114998
Change-Id: I092068c5055d329769986da8d39605cb1bae3a4f
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>