Doc: All overviews list categorization
The \generate list names are added for each categorized section with
some explanation. Here, calling the overviews as
explanations-(name of the section). The idea is to give general terms
instead of specific phrases like 'core' etc, for better understanding.

Allow two warnings for \generatelist commands as the related \group
pages live in modules external to qtdoc and are not loaded when
testing documentation in the CI.

Task-number: QTBUG-115347
Pick-to: 6.5 6.6
Change-Id: Ia8307d42fdb137808ae75cb9bcfeb58de30ba550
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>