Mark the module as free of Q_FOREACH, except where it isn't
The density of Q_FOREACH uses is moderate here, but too high for this
author, unfamiliar with this code, to tackle in a short amount of
time. But they're concentrated in just a few TUs, so pick a different
strategy:

Mark the whole module with QT_NO_FOREACH, to prevent new uses from
creeping in, and whitelist the affected TUs by #undef'ing
QT_NO_FOREACH locally, at the top of each file. For TUs that are part
of a larger executable, this requires these files to be compiled
separately, so add them to NO_PCH_SOURCES (which implies
NO_UNITY_BUILD_SOURCES, too). NB: there are none such TUs in this
module.

Created QTBUG-115830 to keep track of this.

Task-number: QTBUG-115830
Change-Id: I68c9394c04af68cab46480153a5601828565461a
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>