Doc: Restore links to the Plug & Paint example sources
The three examples, 'Plug & Paint' and its plugins, have been moved to
manual tests. This example is tightly integrated to the deployment
documentation for Windows, macOS, and Linux platforms, where it serves
as a tutorial for deploying a Qt application that has plugins of its
own.

While the deployment tools for each platform automates much of the
process, it makes sense to maintain documentation for users who need
to create deployment packages manually.

Add \externalpage links to a source repository for the examples,
remove excess linking to them. In passing, fix a few issues in
deployment and plugin-related docs.

These changes fix around 30 documentation warnings for broken
links to the Plug & Paint example docs.

Fixes: QTBUG-115988
Pick-to: 6.6 6.5
Change-Id: Ifd29ae2a9f16796a520ca7ab4eb6702ccb295326
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>