Update mobile example categories for Qt 6.6
Add mobile example category tags for the chosen examples.
Add for the ones that work best for both Android and iOS.
Remove from the ones that do not work for both (at least yet).

Added, as verified on Android
-calqlatr
Removed (for now), as not verified yet on Android
-todolist
Removed (for now), as not valid anymore nor verified
-qtpurchasing-overview

Fixes: QTBUG-116780
Pick-to: 6.6
Change-Id: I937a90fe54ff87f29cf32627356aa5a65dd51d43
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>