Fix bug in Coffe example causing app stuck when user made a second coffe
The bug was due to the fact that when the coffee was finished brewing,
the state was being set to "initial" which was unnecessary.
So I decided to remove this part of code, which solved the bug.

Task-number: QTBUG-114375
Change-Id: I7c8cbc96330574f0378ea8d1ee3ac2a4732f70e3
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>
(cherry picked from commit 182521e2263f5561d1f28893623db8c01fcebcc9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit ecd23a5f952265effbc28933f084c51903924dab)