Doc: Enable documentation for Same Game example
The documentation for this example - among others - was blacklisted
in commit 012384be6ae. It's unclear what exactly caused this
example to be considered 'broken', but it seems to work now and
is popular, so let's revive it.

Task-number:  QTBUG-119203
Change-Id: Iabdbc2b8b5c84f86e8086a71cfaad4ca8a493f0b
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 2dd01f418286bf0473037906516a96feff660e9d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>