Fix Thermostat Swipe Rooms View in Thermostat example
Thermostat room view is going to the first item, when end of rooms list
is reached and the user wants to go to the next room. Analogous, the
user is sent to the last room when selecting the previous room while
being in the first room.

Fixes: QTBUG-119320
Pick-to: 6.5
Change-Id: If8c18c40ba903ba1b670f122676da6e0eff6c08e
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 26742b74aa9e883510f9a691b08d01fb7823e441)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>