Fix menu icon on Android, fix initial msg not wrapping on small screens
The menu button icon did not appear on Android with the text method, so
added an icon svg to use on all platforms.

The initial label pop up was not wrapping on small screens so added a
height, modified properties to keep the original text size but wrap the
text.

Fixes: QAA-2183
Pick-to: 6.6
Change-Id: I2265e7558dfec839530b5f9b6a7a4129d7b081b2
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Rami Potinkara <rami.potinkara@qt.io>
(cherry picked from commit 39074ca944d946d08a2684816e5eabea4fc400d9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>