Documentviewer demo / Quick3dViewer: Don't re-use quick view
The quick view used to display 3d files in a widgets application was
being re-used. If another viewer is launched and its contents become
visible in the main window scroll area, the quick view's window
container gets deleted, which causes deletion of the quick view as
well.

That makes the application crash, if an .obj file is viewed, followed
by another file type and then .obj again. The latter will crash.

=> Don't re-use the quick view any more.

Fixes: QTBUG-119175
Pick-to: 6.7 6.6
Change-Id: Ieea54f0a3377dfbc9ff3a07e5b99c80b472e5658
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>