Replace the Car-configurator example's car model
Also, other visual improvements are copied from the original QDS
 example.
* SceneEnvironment is replaced by ExtendedSceneEnvironment to enable
 some special effects that are used on scenes like glow effects and
 lens flare effects.
* The opacity of the headlights is changed for the desert scene.
* Light property changes are removed for the studio scene.
* Some unsupported behaviors on eulerRotations are removed.

Also, a bug is fixed for downloading assets
Previously if the unzip command was defined, and the zip file didn't
exist, the process used to give up with downloading assets from the
extracted files. But now, it will try to download the zip file, and
if there is a problem in the process, the rest of the files will be
downloaded from the extracted files.

A simple documentation file is added to the example to enable it in
example list.

Pick-to: 6.7
Change-Id: I4af85d9ea609a5ff20af0c381a096a7c2bbb6c31
Reviewed-by: Kai Köhne <kai.koehne@qt.io>