Android: update the docs and remove outdated or misleading parts
I noticed many parts of the Android docs are still mentioning
outdated or straight wrong information or in other cases, lots
of duplicated information. This is an attempt to remedy that
all at once. What's included in this patch:

* Extend the supported platforms table to include relevant information
like JDK, Gradle and AGP versions (this needs the following patch
46c1414c3bc782f128b24480d5c8367eeb58157b).
* In getting started or building Qt pages, made the dependencies
tools version more dynamic using the macros to have less manual
maintenance.
* Removed unneeded or redundant steps in gettings started and
building Qt pages to have the page more lightweight and straight
to the point.
* In deployment page, re-wrote the page to more efficiently
go through the relevant parts that a user would want to know
about developing, extending and deploying a Qt Android app.
* Added missing info about Gradle, Gradle wrapper and AGP.
* use \annotedlist to collection CMake variables/properties
to dynamically add new values instead of having to do it manually.
* Convert the Android environment variable list to a table
which is better formatted.
* The main Android page had an \annotatedlist with all pages,
but the pages were scattered and it's not easy to know which
are the most relevant pages for someone new, so I moved the most
important pages outside of the annoted list.
* Added CMake snippets or command or prioritized them over
qmake in relevant places.
* Refactored many pages to avoid duplication and add links to the
same information instead.
* In OpenSSL page, prioritize using the pre-built libs over building
them manually, since that's what most users would end up doing.
* In publishing to the store, added parts about using CMake multi-abi
feature.
* In 3rdparty page, added parts about include shared libraries and not
only jar dependencies as it was.
* In platforms notes, removed or moved many parts that were mentioned
elsewhere many times already.
* Moved 'Where to Go from Here' section in main page to end as it makes
sense to go through the Android specific parts and then go elsewhere
only after that.

Task-number: QTBUG-115020
Pick-to: 6.7
Change-Id: I05e73f4a6af216b73b3892b0ec3b1b25174ff0b5
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
Reviewed-by: Rami Potinkara <rami.potinkara@qt.io>