Documentviewer/Q3DViewer: pass correct bounds argument
displayModel.handleBoundsChanged() didn't take an argument and always
handled the bounds of the display model. If a RuntimeLoader was used,
the object was out of bounds.

Make the function take a bounds argument and pass the right bounds to
be handled.

Task-number: QTBUG-119176
Change-Id: I8a9986d9ebea14dea2246f02937fe47bbcf95191
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 85ff73b5ffd8ef56e76745ac457dfeb2ada93592)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>