Lightning Viewer Example: Improve Updating Distance/Time
There was no need to waste time for calculating the distance and time
of the last strike. The calculation should be considered when there's a
need for it, e.g. when the distance/time layer is visible.

Task-number: QTBUG-121123
Change-Id: Ib841502baa5138c91730a308194c6485a28fca1a
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 103e54fe5e0f8864b8402470dd5a39daf4dfa5b8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>