Fix Calqlatr example not scaling properly in landscape mode
Calqlatr example has some number pad buttons that overflow the display
on mobile devices when turning the device to landscape mode.
This happens due to CalculatorButton.qml button component having
hardcoded width and height and contentItem font.pixelSize values, that
do not take into consideration that there might not be enough room
to display these items.

Fixed this by introducing a simple state machine to CalculatorButton.qml
that detects when device is on landscape mode and reduces the
number pad button height and contentItem font.pixelSize enough
so that they'll fit on display.

Button height is set to window.height / 8 in landscape mode to make
the number pad look similar size in whatever display size/density.
There being 7 buttons in a column, so dividing by 8 gives a little
margin there.

Setting the font.pixelSize to 40 seems to work quite well with
displays from 320x480 to 1440x3120.

Fixes: QTBUG-121044
Pick-to: 6.6
Change-Id: I321a037e8f710516ce86b4ff6ad1f94a8b73e179
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>