Lightning Viewer Example: Support LocationPermission
If the permission is granted, the user location will be visible and the
map view can be recentered.
Otherwise, if the user has already denied the permission and triggers
any action that needs the permission, a message dialog will notify the
user that the application needs the location permission, and the
permission is needed to be granted and the application needs to be
restarted.

Additionally, the PositionSource is not active when the permission is
not granted, and the distance and time layer is not able to show the
information until the user allows the application to use the location.

The user's location will be set from QML by PositionSource, then there
is no need to keep the QGeoPositionInfoSource object.

Task-number: QTBUG-121412
Task-number: QTBUG-121123
Change-Id: I78856f8cc8d182aa7b02c1ffaba01a8871e0a0c6
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit 6cc0edc108f5cba0bc407a0f144b79298b9be78d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>