MediaPlayer example: Fix scrolling of track options
The MediaPlayer demo had an issue in the track options list where nested
scrolling caused some options to be difficult to access.

This is fixed by positioning all track options in a Column instead of a
ListView.

Additionally modified the TracksOptions implicit height binding to size
the component based on the amount of options.

Fixes: QTBUG-119285
Pick-to: 6.7 6.6
Change-Id: Iaa1ed2c370118a5c0119b233a59c7f4d8c0c70eb
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>