Add Qt Android Studio Plugin Documentation
Add the page with all the info for the plugin. Add a link in on the
Qt for Android home page. The plugin UI help menu and IntelliJ market
website should link to this as the master.

Task-number: QTBUG-121447
Pick-to: 6.7.0
Change-Id: Icb27903c2c64f0b68e1479a394cd2f150c920a0b
Reviewed-by: Tinja Paavoseppä <tinja.paavoseppa@qt.io>
(cherry picked from commit 60d268cd202f5ee0cadd3ece067607d612323ba0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>