Dice app example: fix menu in landscape mode
Put the column of controls in a flickable and set the scroll-bar
to always on if the UI's height is greater than the screen height.
This way the the user knows they can scroll if they start
the application in a landscape orientation. Should see no change in
on desktops, unless you resize the application window smaller than
the menu height.

Fixes: QTBUG-122444
Pick-to: 6.7.0
Change-Id: Ic49c7b0acdb8f71723925441af5dfbaa8793b67e
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 32fcefb291f092439e317a1b966a4c5a85081e2d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>