Doc: Fix regression in Qt help project configuration
Renaming the 'All Classes' page to 'All C++ Classes' broke the Qt
help (sub)project 'Classes' as it was still referring to the old
page title.

In passing, clean up the rest of the qhp configuration; remove
obsolete 'fake' selector string, and use the 'qmltype' selector
instead of the old, deprecated 'qmlclass'.

Pick-to: 6.7 6.6 6.5
Fixes: QTBUG-120643
Change-Id: I308dc2d4499b9a60ea881cf6bcbf3473dd853421
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>