Lightning Viewer: Update docs to clearly state use of simulated data
Due to the lack of an agreement with the lightning strike data provider,
we will only provide simulated lightning strike data for this demo. This
change removes info in the documentation about the option to forward
real world lightning data, and makes it more clear that the data is
simulated.

Fixes: QTBUG-122907
Pick-to: 6.7
Change-Id: Ibae18922e79393564b3947d813b1285f0358aaf9
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>