Doc: add disclaimer about Linux packages
For some Linux distributions, access to additional repositories may be
required for the package management system to provide the required
development packages for Qt. The packages names may also differ
somewhat among different distributions. Add an example of this for
RedHat and its derivatives to the platform-specific
documentation. Refer users to the documentation for their specific
distribution.

Fixes: QTBUG-122727
Pick-to: 6.7
Change-Id: I8ba2a4245abbf4349ab73a89584741fa2af363f7
Reviewed-by: Liang Qi <liang.qi@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>