CMake: Use qtbase helpers for building the repo
The qtdoc repo need to build the various doc subdirectories instead of
the src subdirectory, so when it was initially ported to CMake, it
didn't use qt_build_repo() like all the other repos, and instead copy
pasted the various pieces of qt_build_repo().

Now that qtbase exposes these as separate functions, we can use them
instead. This mimics the structure qt_build_repo except for the
doc-specific parts.

This also fixes being able to build examples for static builds due to
removing the BUILD_SHARED_LIBS condition, which is not present anymore
in qtbase's qt_build_repo_impl_examples.

Pick-to: 6.7
Change-Id: I0d203b53e785225d594528a5709c65f408e32704
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>