Update What's New 6.7 page to include Qt Quick Timeline feature
Just a mention of the new BlendTrees submodule for Qt Quick Timline.
This module is mostly for enabling tooling like Qt Design Studio and
asset importing in Qt Quick 3D, but it is still new public API so it is
nice to mention it is a new feature.

Change-Id: Ie0dc6605819640a97942522ddffa0192f3f88466
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit b1ada5bd8d064d253465bff6438b8c0c5206f1c8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>