Fix Coffee Machine example initial size on big displays
Fix the issue with the Coffee Machine example application window
filling the whole  virtual desktop screen space, which is wrong.

Set initial application window size to some hardcoded values which
will be used on desktop. On mobile platforms the window size will
be maximized.

Update docs to reflect this change.

Fixes: QTBUG-123414
Pick-to: 6.7
Change-Id: If8323c0c1ada4aa5230aa94cf624b2a0ff859b5e
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>