Set up InAppPurchaseBackend for platforms other than Android and IOS
The hangman example relies on a backend for in-app purchases.
This backend was not properly set up for platforms other than IOS and
Android. This lead to an invalid memory access when utilizing the
backend which resulted in a crash.

Inside the function that does the backend setup, instead of immediately
returning on other platforms, assign an InAppPurchaseBackend object to
the backend variable. This will allow the application to keep running,
and warn the user that InAppPurchaseBackend is not implemented on
the platform.

Fixes: QTBUG-124146
Pick-to: 6.2
Change-Id: I77f64849bf16132f8380fd5b25db89098f86e2e5
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 09eeab6ef42d4f4ee66b19fa3e076d94b99837ad)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d77e15236446db81488032ad9a584fc62d14621f)