Document the -submodules configure option
The -submodules can be easier to use than -skip when trying to reduce
configure and build times because it focuses on the needed submodules
instead of the submodules that you don't need. The documentation should
therefore mention this configure option.

This patch also improves descriptions of the terms submodule, top-level
source directories, and Qt modules, and puts them in context.

Change-Id: I59dbaa9c2dc08360c2ad62efd0c11b03c1595a79
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>