Add a 'Getting Qt Sources from the Git repository' section
This patch adds a documentation page that can be used by new users who
want to obtain the Qt sources from Git.

Building Qt from Git is already documented in Qt Wiki, but because it is
attempting to cover all Qt versions, it is hard to keep it accurate for
the every Qt version, especially after the new -init-repository
configure argument was added.

Documenting the basic steps in the official Qt documentation can
therefore be valuable as a complement to the Qt Wiki.

Pick-to: 6.7
Change-Id: I7e7cbc6568e2128290e48b5a959af87f69c2fa0e
Reviewed-by: Kai Köhne <kai.koehne@qt.io>