Fix Coffee Machine example coffee cup image
Coffee Machine coffee cup image is broken with Qt 6.7.0 for unknown
reasons, more descriptive explanation in the linked bug ticket.

Modified the broken dark_cup.svgz image to expose the hidden
parts of the image.

Before Qt 6.7, Qt only supported SVG Tiny 1.2 standard. From Qt 6.7, we added some of the most used features in SVG 1.1. The file in this example uses <filter> element which was not support before, and since it was not supported it was never parsed. In Qt 6.7, this <filter> element is parsed. However, we only support a subset of the filter primitives that can be used inside the <filter> element, and since this file uses an unsupported filter primitive <feBlend>, it causes an issue.

Simply removing the reference to the <filter> element from the Svg file, fixes the issue.

Fixes: QTBUG-124707
Pick-to: 6.7.1
Change-Id: I62391fa4fc320f440045958e0ef0dc2eef69d21c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 36e202566776d4708fbbeafd84b8d0b86ceeeeaf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>