Remove not needed leftover file from Coffee Machine example
Removed Cup.qml from Coffee Machine example. The file was not supposed
to be there after the example re-design as it is supposed to be
replaced by the Cup.ui.qml but apparently it was not fully removed and
was left in the file listing of coffeemachine.qrc.

With Cmake the file has not been visible in Qt Creator or Squish
but with qmake its visible in Qt Creator
and exposed to Squish as well.

Affects testing the example because the same object has different
object names in Squish between Cmake and qmake.

Task-number: QTBUG-123440
Pick-to: 6.7 6.7.1
Change-Id: I82c2a315cd8af9459e5d5caecb3a3efe8a1e5eff
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Jani Korteniemi <jani.korteniemi@qt.io>