Fix small mistakes in QtTAS documentation
Fix the wrong path to the example - this is not a major problem as the
screenshot below already shows the correct example path.

This was discussed in the earlier code review but we still managed to
get in wrong.

Also fix two small style issues with capitalization.

Change-Id: I2e825022202c4e250ad8b5ed34c35ad6d68c8356
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
(cherry picked from commit e08c27ff1b31abc24b998d3aa4529b188424a7b5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>