Doc: Refresh 'All Qt Modules' page
Use the new macros to output a grid of items, instead of using \table
commands.

Simplify descriptions for modules, and drop links from descriptions as
the grid element does not support nested links (the entire grid item
acts as the link).

Move target platform information from a dedicated column into the
description when needed. In most cases, this information can be
omitted entirely, as the platform support can be expected to
live in the module's documentation.

Pick-to: 6.8
Task-number: QTWEBSITE-1144
Change-Id: I59f86496a34e84a66098e4aa8b23ef9c213fc838
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>