Clarify macdeployqt behavior for image format and icon engine plugins
The correctness of this behavior can be argued, but this has been the
behavior for quite some time now, so let's document that as a start.

Once we plumb the CMake qt_import_plugins API to deployment we can
be more selective in how we deploy plugins.

Fixes: QTBUG-55199
Pick-to: 6.7
Change-Id: I9b17eb047af27277c43a4167011d4ce3d0502303
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit c9d5c228f6799ec085e1f91ce284952e32ffcb39)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>