clock demo: use qt_add_qml_module
Make the clocks demo use qt_add_qml_module instead of qt_add_resources.
Also put the qml files in the same folder where the qt_add_qml_module
happens to avoid having two qml modules (one with clocks.qml and another
with Clock.qml), and move the images to an images folder.
Rename the executable to clocksexample to avoid clashes with the newly
added QML module clocks.

Fix the uris to the images to reflect the new path in the qrc.

Use qt_standard_project_setup().

Use QQuickView::loadFromModule() in the main.cpp and add a qmldir in the
.qrc to make loadFromModule() work when using qmake.

Also update the documentation.

Pick-to: 6.8
Task-number: QTBUG-100100
Change-Id: I16274f041f6d16acc5910d9eaef645234dc6f34f
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>